fix: move ssr config to modelence.config.ts
