+++ packages/modelence/src/app/index.test.ts
@@ -244,6 +244,13 @@ vi.doMock('../data/resolveStores', () => ({
   resolveStores: mockResolveStores,
 }));
 
+vi.doMock('./modelenceConfig', () => ({
+  loadModelenceConfig: vi.fn(async () => ({
+    serverDir: 'src/server',
+    serverEntry: 'app.ts',
+  })),
+}));
+
 const { startApp } = await import('./index');
 
 // Helper to create a test module
